Complete delete functionality with backend
